+++ src/pages/api/node/[trpc].ts
@@ -1,8 +1,8 @@
 import { createNextApiHandler } from '@trpc/server/adapters/next';
 
 import { env } from '~/configs/env.mjs';
-import { appNodeRouter } from '~/server/api/root';
-import { createTRPCContext } from '~/server/api/trpc';
+import { appNodeRouter } from '~/server/api/routers/root';
+import { createTRPCContext } from '~/server/utils/trpc';
 
 export const config = {
   runtime: 'nodejs',
