+++ src/server/api/routers/edge/index.ts
@@ -1,0 +1,5 @@
+import { exampleEdgeRouter } from '~/server/api/routers/edge/example';
+
+export const edgeRouters = {
+  example: exampleEdgeRouter,
+};
