+++ src/server/api/routers/node/index.ts
@@ -1,0 +1,5 @@
+import { exampleNodeRouter } from '~/server/api/routers/node/example';
+
+export const nodeRouters = {
+  example: exampleNodeRouter,
+};
